chore(fromEvent): move types test to dtslint & remove comments
